Use pybind for rd_file
This fixes a bug with the report_steps and num_report_steps
methods, which would throw KeyError on files without "SEQNUM".
It now returns []/0 as documented.

The `has_kw` and `report_dates` functions had imprecise docstrings,
which were updated. The module docstring referenced rd_file.c which
no longer exists: so this internal detail was removed from the
docstring.

+++ lib/include/resdata/rd_file.hpp
@@ -41,7 +41,7 @@ struct File {
         return global_view->num_named_kw(kw);
     }
     /** Will return the ith occurence of @kw the File. */
-    [[nodiscard]] rd_kw_type *get_kw(const std::string &kw, int ith) const {
+    [[nodiscard]] rd_kw_type *get_kw(const std::string &kw, size_t ith) const {
         return global_view->get_kw(kw, ith);
     }
     /** The total number of rd_kws in the File. */
